requirements.txt
click==6.6
Flask==0.12
itsdangerous==0.24
Jinja2==2.8
jsonify==0.5
MarkupSafe==0.23
phue==0.9
pkg-resources==0.0.0
pprint==0.1
PyYAML==3.12
Werkzeug==0.11.11
